Move public/js/queries to workspace/templates
