+++ mod/workspace/templates/layer_extent.js
