--- public/js/queries/count_locations.js
@@ -1,4 +1,0 @@
-module.exports = `
-  SELECT count(*)
-  FROM \${table}
-  WHERE true \${filter};`
