--- public/js/queries/nnearest_locations.sql
@@ -1,4 +1,0 @@
-SELECT *
-  FROM ${table}
-  WHERE true ${filter}
-  ORDER BY ST_Point(%{x},%{y}) <#> ${geom} LIMIT ${n};
